reorganized to have all data management here, added new page functionality
